// src/app/interfaces/about.ts
export interface About {
  id?: string;
  title?: string;
  titleSpan?: string;
  subtitle?: string;
  desc?: string;
  para1?: string;
  para2?: string;
  para3?: string;
  btnText?: string;
  btnLink?: string;
  imageUrl?: string;
  titleTwo?: string;
  titleTwoSpan?: string;
  qualifications?: any;
}

export interface AboutCta {
  id?: string;
  title?: string;
  titleSpan?: string;
  desc?: string;
  para?: string;
  btnText?: string;
  btnLink?: string;
}

export interface Qualification {
  id?: string;
  title?: string;
}

export interface Gallery {
  id?: string;
  imageUrl?: string;
}